+++ src/Edgar/Legacy/Core/LayoutGenerators/DungeonGenerator/DungeonGenerator.cs
@@ -57,7 +57,7 @@ namespace Edgar.Legacy.Core.LayoutGenerators.DungeonGenerator
         /// <summary>
         /// Total time to generate a level.
         /// </summary>
-        public double TimeTotal => generator.TimeTotal;
+        public long TimeTotal => generator.TimeTotal;
 
         /// <summary>
         /// Number of iterations needed to generate the last level.
